# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MediQore REST API
  version: 0.6.0
  description: |
    REST API for the MediQore LHW app and the supervisor/admin portal (roadmap P0-3, P0-6; Phase 1: M1 FE-1–3, M2, M3 FE-2, M10 base).

    Conventions (see CLAUDE.md):
    - Every route is under `/api/v1` and exchanges JSON.
    - Every request body is validated with Joi before any database call.
    - Errors always use the `Error` shape below.
    - Authentication is a short-lived JWT access token plus a refresh token (M1 FE-2). Every route
      requires the access token by default; public routes such as /health and /auth/* set `security: []`.
    - The app signs in with its installation ID. A phone that has not been approved answers
      `otp_required` until the user types the one-time code an admin or supervisor issued for it
      (decision 0002, channel `admin_issued`). LHWs can only sign in through the app.
    - Failed sign-ins are rate-limited per client address and username (429 TOO_MANY_ATTEMPTS with Retry-After).
    - HTTPS only in production: plain HTTP is refused with 403 HTTPS_REQUIRED.
    - Records are scoped by area: an LHW sees only her area, a supervisor only their assigned areas.
    - Each route checks a permission, and each permission belongs to fixed roles (M10 FE-3, see GET /admin/roles).

    Document every new route here in the same pull request that adds it.
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
...
security:
  - bearerAuth: []
...
paths:
...
  /admin/areas:
    get:
      tags: [Admin]
      summary: Every area with its Union Council, tehsil and district
      operationId: listAreas
      responses:
        '200':
          description: Areas, ordered by district, tehsil, Union Council and name.
          content:
            application/json:
              schema:
                type: object
                required: [areas]
                properties:
                  areas:
                    type: array
                    items:
                      $ref: '#/components/schemas/AreaOption'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
  /admin/lhws:
    get:
      tags: [Admin]
      summary: List LHWs
      operationId: listLhws
      parameters:
        - name: search
          in: query
          description: Part of the name or LHW ID.
          schema:
            type: string
        - name: areaId
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, all]
            default: all
      responses:
        '200':
          description: LHWs, ordered by LHW ID.
          content:
            application/json:
              schema:
                type: object
                required: [lhws]
                properties:
                  lhws:
                    type: array
                    items:
                      $ref: '#/components/schemas/Lhw'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
    post:
      tags: [Admin]
      summary: Create an LHW (M1 FE-1)
      description: The system issues the next LHW ID (LHW-00001, ...), which is also the username, and a random password. The password is returned once and stored only as a hash.
      operationId: createLhw
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fullName, areaId]
              properties:
                fullName:
                  type: string
                  minLength: 2
                  maxLength: 100
                areaId:
                  type: string
                  format: uuid
                phone:
                  type: string
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LhwWithCredentials'
        '400':
          description: VALIDATION_ERROR or UNKNOWN_AREA.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
  /admin/lhws/{id}:
    patch:
      tags: [Admin]
      summary: Rename an LHW, change the phone number or reassign the area (M1 FE-3)
      operationId: updateLhw
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              minProperties: 1
              properties:
                fullName:
                  type: string
                areaId:
                  type: string
                  format: uuid
                phone:
                  type: string
      responses:
        '200':
          $ref: '#/components/responses/LhwResult'
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /admin/lhws/{id}/deactivate:
    post:
      tags: [Admin]
      summary: Deactivate an LHW; refresh tokens are revoked and the next sync is refused (M1 FE-3)
      operationId: deactivateLhw
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          $ref: '#/components/responses/LhwResult'
        '404':
          $ref: '#/components/responses/Error'
  /admin/lhws/{id}/activate:
    post:
      tags: [Admin]
      summary: Reactivate an LHW (M1 FE-3)
      operationId: activateLhw
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          $ref: '#/components/responses/LhwResult'
        '404':
          $ref: '#/components/responses/Error'
...
  /admin/staff:
    get:
      tags: [Admin]
      summary: List supervisor and admin accounts (M10 FE-3)
      operationId: listStaff
      parameters:
        - name: search
          in: query
          description: Part of the name or username.
          schema:
            type: string
        - name: role
          in: query
          schema:
            type: string
            enum: [supervisor, admin]
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, all]
            default: all
      responses:
        '200':
          description: Accounts.
          content:
            application/json:
              schema:
                type: object
                required: [staff]
                properties:
                  staff:
                    type: array
                    items:
                      $ref: '#/components/schemas/Staff'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
    post:
      tags: [Admin]
      summary: Create a supervisor or admin account
      description: The admin chooses the username; the password is random, returned once and stored only as a hash. A supervisor needs at least one area.
      operationId: createStaff
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role, username, fullName]
              properties:
                role:
                  type: string
                  enum: [supervisor, admin]
                username:
                  type: string
                  pattern: '^[a-z][a-z0-9._-]{2,31}$'
                  description: Stored in lower case.
                fullName:
                  type: string
                  minLength: 2
                  maxLength: 100
                phone:
                  type: string
                areaIds:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StaffWithCredentials'
        '400':
          description: VALIDATION_ERROR, AREAS_REQUIRED or UNKNOWN_AREA.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: USERNAME_TAKEN.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /admin/staff/{id}:
    patch:
      tags: [Admin]
      summary: Change the name, phone, role or areas of a supervisor or admin
      description: A role change signs the account out. An admin cannot change their own role (OWN_ACCOUNT), and the last active admin stays an admin (LAST_ADMIN). Areas are replaced by the list given.
      operationId: updateStaff
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              minProperties: 1
              properties:
                role:
                  type: string
                  enum: [supervisor, admin]
                fullName:
                  type: string
                phone:
                  type: string
                areaIds:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        '200':
          $ref: '#/components/responses/StaffResult'
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /admin/staff/{id}/deactivate:
    post:
      tags: [Admin]
      summary: Deactivate a supervisor or admin; their refresh tokens are revoked
      description: Not your own account (OWN_ACCOUNT), and not the last active admin (LAST_ADMIN).
      operationId: deactivateStaff
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          $ref: '#/components/responses/StaffResult'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /admin/staff/{id}/activate:
    post:
      tags: [Admin]
      summary: Reactivate a supervisor or admin
      operationId: activateStaff
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          $ref: '#/components/responses/StaffResult'
        '404':
          $ref: '#/components/responses/Error'
  /admin/staff/{id}/reset-password:
    post:
      tags: [Admin]
      summary: Reset a supervisor's or admin's password; returned once, refresh tokens revoked
      operationId: resetStaffPassword
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: The new credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StaffWithCredentials'
        '404':
          $ref: '#/components/responses/Error'
  /admin/roles:
    get:
      tags: [Admin]
      summary: The fixed roles and what each may do (M10 FE-3)
      description: The same permission list every route checks.
      operationId: listRoles
      responses:
        '200':
          description: Roles and permissions.
          content:
            application/json:
              schema:
                type: object
                required: [roles, permissions]
                properties:
                  roles:
                    type: array
                    items:
                      type: object
                      required: [id, label, description]
                      properties:
                        id:
                          type: string
                          enum: [lhw, supervisor, admin]
                        label:
                          type: string
                        description:
                          type: string
                  permissions:
                    type: array
                    items:
                      type: object
                      required: [id, label, roles]
                      properties:
                        id:
                          type: string
                          example: audit.view
                        label:
                          type: string
                        roles:
                          type: array
                          items:
                            type: string
                            enum: [lhw, supervisor, admin]
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
  /admin/geography:
    get:
      tags: [Admin]
      summary: The district > tehsil > Union Council > area tree (M10 FE-3)
      description: Deleted units are left out. Each area says how many LHWs and supervisors it has.
      operationId: geographyTree
      responses:
        '200':
          description: The tree, each level ordered by name.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeographyTree'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
  /admin/geography/{level}:
    post:
      tags: [Admin]
      summary: Add a district, tehsil, Union Council or area
      description: Below district level, parentId is the unit above. Names are unique among siblings (case-insensitive); a name that belonged to a deleted sibling brings that unit back with its ID.
      operationId: createUnit
      parameters:
        - $ref: '#/components/parameters/Level'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  minLength: 2
                  maxLength: 100
                parentId:
                  type: string
                  format: uuid
      responses:
        '201':
          $ref: '#/components/responses/UnitResult'
        '400':
          description: VALIDATION_ERROR or UNKNOWN_PARENT.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: DUPLICATE_NAME.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /admin/geography/{level}/{id}:
    patch:
      tags: [Admin]
      summary: Rename a unit
      operationId: renameUnit
      parameters:
        - $ref: '#/components/parameters/Level'
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  minLength: 2
                  maxLength: 100
      responses:
        '200':
          $ref: '#/components/responses/UnitResult'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
    delete:
      tags: [Admin]
      summary: Delete a unit (soft delete), only when nothing uses it
      description: Refused with 409 IN_USE while it has children, accounts, facilities or records; details.uses says what.
      operationId: deleteUnit
      parameters:
        - $ref: '#/components/parameters/Level'
        - $ref: '#/components/parameters/Id'
      responses:
        '204':
          description: Deleted.
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
...
  /admin/audit:
    get:
      tags: [Audit]
      summary: The audit log, newest first (M10 FE-3)
      description: Every create, edit, delete, referral, alert, login and sync conflict, with the user and time. Dates are whole days, Pakistan time. Page back with `before` = `nextBefore`.
      operationId: listAudit
      parameters:
        - name: user
          in: query
          description: Part of the username or name of who did it.
          schema:
            type: string
        - name: action
          in: query
          schema:
            type: string
            enum: [create, edit, delete, referral, alert, login, sync_conflict]
        - name: entityType
          in: query
          description: The record type (table name), for example users or visits.
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
        - name: before
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: A page of entries, with the filter choices.
          content:
            application/json:
              schema:
                type: object
                required: [entries, nextBefore, actions, entityTypes]
                properties:
                  entries:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditEntry'
                  nextBefore:
                    type: [integer, 'null']
                    description: Pass as `before` for the next page; null on the last page.
                  actions:
                    type: array
                    items:
                      type: string
                  entityTypes:
                    type: array
                    items:
                      type: string
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Level:
      name: level
      in: path
      required: true
      schema:
        type: string
        enum: [districts, tehsils, union-councils, areas]
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Error:
      description: Error response in the standard shape.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    StaffResult:
      description: The account after the change.
      content:
        application/json:
          schema:
            type: object
            required: [staff]
            properties:
              staff:
                $ref: '#/components/schemas/Staff'
    UnitResult:
      description: The unit.
      content:
        application/json:
          schema:
            type: object
            required: [unit]
            properties:
              unit:
                type: object
                required: [id, name, parentId]
                properties:
                  id:
                    type: string
                    format: uuid
                  name:
                    type: string
                  parentId:
                    type: [string, 'null']
                    format: uuid
...
    LhwResult:
      description: The LHW after the change.
      content:
        application/json:
          schema:
            type: object
            required: [lhw]
            properties:
              lhw:
                $ref: '#/components/schemas/Lhw'
  schemas:
...
    AreaOption:
      type: object
      required: [id, name, unionCouncil, tehsil, district]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        unionCouncil:
          type: string
        tehsil:
          type: string
        district:
          type: string
    NamedRef:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
    Lhw:
      type: object
      required: [id, lhwCode, username, fullName, isActive, area, devices]
      properties:
        id:
          type: string
          format: uuid
        lhwCode:
          type: string
          example: LHW-00001
        username:
          type: string
        fullName:
          type: string
        phone:
          type: [string, 'null']
        isActive:
          type: boolean
        lastLoginAt:
          type: [string, 'null']
          format: date-time
        createdAt:
          type: string
          format: date-time
        area:
          type: object
          required: [id, name, unionCouncil, tehsil, district]
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
            unionCouncil:
              $ref: '#/components/schemas/NamedRef'
            tehsil:
              $ref: '#/components/schemas/NamedRef'
            district:
              $ref: '#/components/schemas/NamedRef'
        devices:
          type: object
          required: [approved, pending]
          properties:
            approved:
              type: integer
            pending:
              type: integer
    LhwWithCredentials:
      type: object
      required: [lhw, credentials]
      properties:
        lhw:
          $ref: '#/components/schemas/Lhw'
        credentials:
          type: object
          required: [username, password]
          description: Shown to the admin once; never stored in plain text.
          properties:
            username:
              type: string
            password:
              type: string
    Staff:
      type: object
      required: [id, role, username, fullName, isActive, areas]
      properties:
        id:
          type: string
          format: uuid
        role:
          type: string
          enum: [supervisor, admin]
        username:
          type: string
        fullName:
          type: string
        phone:
          type: [string, 'null']
        isActive:
          type: boolean
        lastLoginAt:
          type: [string, 'null']
          format: date-time
        createdAt:
          type: string
          format: date-time
        areas:
          type: array
          description: A supervisor's areas; empty for an admin.
          items:
            $ref: '#/components/schemas/AreaOption'
    StaffWithCredentials:
      type: object
      required: [staff, credentials]
      properties:
        staff:
          $ref: '#/components/schemas/Staff'
        credentials:
          type: object
          required: [username, password]
          description: Shown to the admin once; never stored in plain text.
          properties:
            username:
              type: string
            password:
              type: string
    GeographyTree:
      type: object
      required: [districts]
      properties:
        districts:
          type: array
          items:
            type: object
            required: [id, name, tehsils]
            properties:
              id:
                type: string
                format: uuid
              name:
                type: string
              tehsils:
                type: array
                items:
                  type: object
                  required: [id, name, unionCouncils]
                  properties:
                    id:
                      type: string
                      format: uuid
                    name:
                      type: string
                    unionCouncils:
                      type: array
                      items:
                        type: object
                        required: [id, name, areas]
                        properties:
                          id:
                            type: string
                            format: uuid
                          name:
                            type: string
                          areas:
                            type: array
                            items:
                              type: object
                              required: [id, name, lhws, supervisors]
                              properties:
                                id:
                                  type: string
                                  format: uuid
                                name:
                                  type: string
                                lhws:
                                  type: integer
                                supervisors:
                                  type: integer
...
    AuditEntry:
      type: object
      required: [id, occurredAt, action, details]
      properties:
        id:
          type: integer
        occurredAt:
          type: string
          format: date-time
        action:
          type: string
          enum: [create, edit, delete, referral, alert, login, sync_conflict]
        entityType:
          type: [string, 'null']
        entityId:
          type: [string, 'null']
          format: uuid
        entityLabel:
          type: [string, 'null']
          description: For accounts, the username.
        deviceId:
          type: [string, 'null']
          format: uuid
        details:
          type: object
        user:
          description: Who did it; null when unknown, for example a failed sign-in.
          oneOf:
            - type: 'null'
            - type: object
              required: [id, username, fullName, role]
              properties:
                id:
                  type: string
                  format: uuid
                username:
                  type: string
                fullName:
                  type: string
                role:
                  type: string
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable code.
              enum:
                - VALIDATION_ERROR
                - INVALID_JSON
                - PAYLOAD_TOO_LARGE
                - NOT_FOUND
                - UNAUTHORIZED
                - INVALID_CREDENTIALS
                - FORBIDDEN
                - NO_AREA
                - DEVICE_NOT_ALLOWED
                - DEVICE_REQUIRED
                - ACCOUNT_INACTIVE
                - OTP_NOT_ISSUED
                - OTP_INVALID
                - OTP_LOCKED
                - INVALID_REFRESH_TOKEN
                - TOO_MANY_ATTEMPTS
                - HTTPS_REQUIRED
                - UNKNOWN_AREA
                - ALREADY_RESOLVED
                - UNKNOWN_PARENT
                - DUPLICATE_NAME
                - IN_USE
                - USERNAME_TAKEN
                - AREAS_REQUIRED
                - OWN_ACCOUNT
                - LAST_ADMIN
                - UNKNOWN_DISTRICT
                - AREA_NOT_IN_DISTRICT
                - INTERNAL_ERROR
            message:
              type: string
            details:
              description: For VALIDATION_ERROR, one entry per problem. For IN_USE, `uses` lists what still uses the record.
              oneOf:
                - type: array
                  items:
                    type: object
                    required: [in, path, message]
                    properties:
                      in:
                        type: string
                        enum: [params, query, body]
                      path:
                        type: string
                      message:
                        type: string
                - type: object
                  properties:
                    uses:
                      type: array
                      items:
                        type: string